move params into keeper.go
