+++ x/gamm/keeper/keeper.go
@@ -70,3 +70,14 @@ func (k *Keeper) SetHooks(gh types.GammHooks) *Keeper {
 
 	return k
 }
+
+// GetParams returns the total set params.
+func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
+	k.paramSpace.GetParamSet(ctx, &params)
+	return params
+}
+
+// SetParams sets the total set of params.
+func (k Keeper) setParams(ctx sdk.Context, params types.Params) {
+	k.paramSpace.SetParamSet(ctx, &params)
+}
