--- x/gamm/keeper/params.go
@@ -1,18 +1,0 @@
-package keeper
-
-import (
-	"github.com/osmosis-labs/osmosis/v10/x/gamm/types"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-// GetParams returns the total set params.
-func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
-	k.paramSpace.GetParamSet(ctx, &params)
-	return params
-}
-
-// SetParams sets the total set of params.
-func (k Keeper) setParams(ctx sdk.Context, params types.Params) {
-	k.paramSpace.SetParamSet(ctx, &params)
-}
